refactor(itunes): move TrackProvisioner into itunesservice (M1 step 1)
First real sub-component move of Phase 2 M1. Relocates the iTunes
track provisioner — a ~145-line leaf module with no HTTP surface —
from internal/server/ into the new internal/itunes/service/ package
as a proper struct-with-methods.

## What moves
- internal/server/itunes_track_provisioner.go (deleted)
- internal/itunes/service/track_provisioner.go (new, 180 lines)
  - TrackProvisioner struct with {store, enqueuer, cfg} fields
  - Provision / ProvisionAll / SetEnqueuer methods (was free funcs)
  - bookAuthor is now a method on TrackProvisioner
  - linuxToWindowsPath / kindFromExt stay as package-local helpers
- internal/itunes/service/enqueuer.go (new)
  - Canonical Enqueuer interface now lives in itunesservice

## Bridging the Enqueuer type
Server-package code still references `server.Enqueuer` in many places
(struct fields, constructor params). To avoid churning every site in
this PR, internal/server/writeback_enqueuer.go is now a type alias:
`type Enqueuer = itunesservice.Enqueuer`. Go aliases are literally
the same type — `*WriteBackBatcher` satisfies both names, so no
conversion is needed at call sites.

## Config surface grows
internal/itunes/service/config.go gains two fields:
- AutoWriteBack      (from config.AppConfig.ITunesAutoWriteBack)
- ITLWriteBackEnabled (from config.AppConfig.ITLWriteBackEnabled)
The provisioner reads p.cfg.AutoWriteBack as its gate — same check the
old free function did against config.AppConfig, just routed through
the injected Config value (same pattern as P3 for the batcher).

## Service.New is no longer always-disabled
The foundation PR #401 stubbed `server.itunesSvc = NewDisabled()`.
This PR flips that to conditional `itunesservice.New(deps)` with
Enabled=true so Service.Provisioner is populated. Disabled fallback
kept for future test paths. ImportService now holds a
*TrackProvisioner reference (via SetTrackProvisioner) instead of a
raw Enqueuer, and calls svc.Provisioner.ProvisionAll(book) at the
one place that was ProvisionITLTracksForBook.

## Transitional batcher-wiring
The batcher itself still lives in the server package and is wired
into the provisioner after construction via SetEnqueuer. When the
batcher moves (M1 step 3), the wiring collapses: Service.New holds
the batcher directly and passes it to newTrackProvisioner at
construction time — SetEnqueuer becomes unused.

## Pre-work dependencies consumed
- P1 (#403): Enqueuer interface — promoted from server alias to
  canonical location here
- P7 (#409): indirection patterns validated (factory var for
  organizer, same shape works here for sub-components)

Plan: docs/superpowers/specs/2026-04-18-itunes-service-extraction-v2-notes.md
